EOS-19418: Skip meta-data lookup, upon receiving pver from application
Added flag in object entity as "en_flags" to know that particular application has
capability to store meta-data, If M0_EOF_META flag is set that means application
can store pver, LID in it's meta-data and application will send this
in m0_entity_open() while reading object so the meta-data lookup can be skipped.
In case of motr application where meta-data can not be store we have to go for
meta-data lookup to get pool version.

Signed-off-by: Yatin Mahajan <yatin.mahajan@seagate.com>